Measure amplitude envelope correlation

env_cor Measure amplitude envelope correlation in signals referenced in a extended selection table.

env_cor(X, parallel = 1, pb = TRUE, method = 1, cor.method = "pearson",
ssmooth = NULL, msmooth = NULL, output = "est")

Arguments

X
object of class 'selection_table', 'extended_selection_table' created by the function selection_table from the warbleR package.

parallel
Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing). If NULL (default) then the current working directory is used.

pb
Logical argument to control if progress bar is shown. Default is TRUE. It can also be set globally using the 'pb' option (see warbleR_options).

method
Numeric vector of length 1 to indicate the 'experimental design' for measuring envelope correlation. Two methods are available:

1: compare all signals with their counterpart that was recorded at the closest distance to source (e.g. compare a signal recorded at 5m, 10m and 15m with its counterpart recorded at 1m). This is the default method.

2: compare all signals with their counterpart recorded at the distance inmediately before (e.g. a signal recorded at 10m compared with the same signal recorded at 5m, then signal recorded at 15m compared with same signal recorded at 10m and so on).

cor.method
Character string indicating the correlation coefficient to be applied ("pearson", "spearman", or "kendall", see cor).

ssmooth
Numeric vector of length 1 determining the length of the sliding window used for a sum smooth for amplitude envelope calculation (used internally by env).

msmooth
Numeric vector of length 2 to smooth the amplitude envelope with a mean sliding window for amplitude envelope calculation. The first element is the window length (in number of amplitude values) and the second one the window overlap (used internally by env).

output
Character vector of length 1 to determine if an extended selection table ('est', default) or a data frame ('data.frame') is returned.

Value

Data frame or extended selection table (depending on 'output' argument) similar to input data, but also includes a new column with the amplitude envelope correlation coefficients.

Details

The correlation of amplitude envelopes is intended to measure the distorcion of signals in the time domain. The goal of the function is to measure the envelope correlation on signals in which a master playback has been re-recorded at different distances. The 'signal.id' column must be used to tell the function to only compare signals belonging to the same category (e.g. song-types). The function will then compared each signal type to its reference. Two methods for calculating excess attenuation are provided (see 'method' argument).

Examples

{
# First set temporary folder
# setwd(tempdir())

data("playback_est")

# method 1
env_cor(X = playback_est)

# method 2
env_cor(X = playback_est, method = 2)
}

#> calculating amplitude envelopes (step 1 of 2):
#>   |                                                  | 0 % ~calculating    |+++                                               | 5 % ~00s            |+++++                                             | 10% ~00s            |++++++++                                          | 15% ~00s            |++++++++++                                        | 20% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++++                                | 35% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> calculating envelope cross-correlations (step 2 of 2):
#>   |                                                  | 0 % ~calculating    |+++                                               | 5 % ~00s            |+++++                                             | 10% ~00s            |++++++++                                          | 15% ~00s            |++++++++++                                        | 20% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++++                                | 35% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> calculating amplitude envelopes (step 1 of 2):
#>   |                                                  | 0 % ~calculating    |+++                                               | 5 % ~00s            |+++++                                             | 10% ~00s            |++++++++                                          | 15% ~00s            |++++++++++                                        | 20% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++++                                | 35% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> calculating envelope cross-correlations (step 2 of 2):
#>   |                                                  | 0 % ~calculating    |+++                                               | 5 % ~00s            |+++++                                             | 10% ~00s            |++++++++                                          | 15% ~00s            |++++++++++                                        | 20% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++++                                | 35% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  

#> object of class 'extended_selection_table' 
#>  contains a selection table data frame with 20 rows and 9 columns: 
#>                                      sound.files selec start       end
#> 22410 Playback_sound_file_LBH_degradation.wav_28     1   0.1 0.2443537
#> 12488             2W.5m.2019.7.1.10.10.wav_12488     1   0.1 0.2435374
#> 12052            2W.10m.2019.7.1.10.20.wav_12052     1   0.1 0.2443537
#> 12268            2W.15m.2019.7.1.10.30.wav_12268     1   0.1 0.2463039
#> 22393 Playback_sound_file_LBH_degradation.wav_11     1   0.1 0.2824490
#> 12392             2W.5m.2019.7.1.10.10.wav_12392     1   0.1 0.2860544
#>       bottom.freq top.freq signal.id distance   env.cor
#> 22410    2.007155 11.35765    BR1-D4        1        NA
#> 12488    1.891390 11.24189    BR1-D4        5 0.2862830
#> 12052    2.007155 11.35765    BR1-D4       10 0.3792296
#> 12268    1.833508 11.18401    BR1-D4       15 0.3084996
#> 22393    2.110482 11.46098    CCE-I2        1        NA
#> 12392    2.776771 12.12727    CCE-I2        5 0.4149716
#> ... and 14 more rows 
#> 20 wave objects (as attributes): 
#> [1] "Playback_sound_file_LBH_degradation.wav_6" 
#> [2] "Playback_sound_file_LBH_degradation.wav_11"
#> [3] "Playback_sound_file_LBH_degradation.wav_28"
#> [4] "Playback_sound_file_LBH_degradation.wav_48"
#> [5] "Playback_sound_file_LBH_degradation.wav_54"
#> [6] "2W.10m.2019.7.1.10.20.wav_12052"           
#> ... and 14 more 
#> and a data frame (check.results) generated by checkres() (as attribute) 
#> the selection table was created by element (see 'class_extended_selection_table')