Plot spectrograms to check test sound files alignment

plot_aligned_sounds() plots spectrograms to visually inspect alignment precision on test sound files.

Usage

plot_aligned_sounds(
  X,
  hop.size = getOption("hop.size", 11.6),
  wl = getOption("wl", NULL),
  ovlp = getOption("ovlp", 50),
  path = getOption("sound.files.path", "."),
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  collevels = seq(-120, 0, 5),
  palette = viridis::viridis,
  duration = 2,
  mar = 0.2,
  dest.path = getOption("dest.path", "."),
  flim = NULL,
  col = "white",
  width = 7,
  height = 4,
  res = 100,
  label = TRUE,
  fast.spec = FALSE,
  srt = 0,
  cex = 1,
  ...
)

Arguments

X

Object of class data.frame, selection_table, or extended_selection_table (the last 2 classes are created by warbleR::selection_table() from the warbleR package) with the test sound files' annotations. Must contain the following columns: 1) sound.files: name of the .wav files, 2) selec: unique selection identifier (within a sound file), 3) start: start time and 4) end: end time of selections, 5) bottom.freq: low frequency for bandpass, 6) top.freq: high frequency for bandpass, and 7) sound.id: ID of sounds used to identify counterparts across distances. Each sound must have a unique ID within a distance.

hop.size

Numeric vector of length 1 specifying the time window duration (in ms). Default 11.6 ms, which is equivalent to 512 wl for a 44.1 kHz sampling rate. Ignored if wl is supplied. Can be set globally for the current R session via the "hop.size" option (see options()).

wl

Numeric vector of length 1 specifying the window length of the spectrogram. Default NULL. Ignored if bp = NULL. If supplied, hop.size is ignored.

ovlp

Numeric vector of length 1 specifying the percentage of overlap between two consecutive windows, as in seewave::spectro(). Default 0. Can be set globally for the current R session via the "ovlp" option (see options()).

path

Character string containing the directory path where the sound files are found. Only needed when X is not an extended selection table. If not supplied the current working directory is used. Can be set globally for the current R session via the "sound.files.path" option (see options()).

cores

Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default 1 (i.e. no parallel computing). Can be set globally for the current R session via the "mc.cores" option (see options()).

pb

Logical argument to control if progress bar is shown. Default TRUE. Can be set globally for the current R session via the "pb" option (see options()).

collevels

Numeric vector of length 3. Specifies levels to partition the amplitude range of the spectrogram (in dB). The more levels, the higher the resolution of the spectrogram. Default seq(-40, 0, 1). seq(-115, 0, 1) will produce spectrograms similar to other acoustic analysis software packages.

palette

Color palette function for the spectrogram. Default viridis::viridis(). See seewave::spectro() for more palettes. Palettes such as monitoR::gray.2 may work better when fast.spec = TRUE.

duration

Numeric vector of length 1. Specifies the overall duration of the clip that will be plotted. Notice that only the initial part of the test files is plotted, as this is usually enough to tell the precision of the alignment.

mar

Numeric vector of length 1. Specifies the margins adjacent to the start of the first annotation to be included in the plot.

dest.path

Character string containing the directory path where the image files will be saved. If not supplied the current working directory will be used instead. Can be set globally for the current R session via the "dest.path" option (see options()).

flim

Numeric vector of length 2 indicating the highest and lowest frequency limits (kHz) of the spectrogram, as in seewave::spectro(). Default NULL, which will plot spectrograms in the full frequency range (0 - nyquist frequency).

col

Character string controlling the color of lines and sound ID labels.

width

Numeric vector of length 1. Single value (in inches) indicating the width of the output image files. Default 7.

height

Numeric vector of length 1. Single value (in inches) indicating the height of the output image files. Default 4.

res

Numeric argument of length 1. Controls image resolution. Default 100 (faster), although 300-400 is recommended for publication/presentation quality.

label

Logical to control if labels (from the sound.id column in X) are plotted. Default TRUE.

fast.spec

Logical. If TRUE, the image function is used internally to create spectrograms, which substantially increases performance (much faster), although some options become unavailable, such as collevels (amplitude scale). Default FALSE.

srt

Numeric argument of length 1. The rotation (in degrees) of the sound ID labels. Default 0.

cex

Numeric argument of length 1 controlling the size of sound ID text labels. Default 1.

...

Additional arguments to be passed to the internal spectrogram-creating function for customizing graphical output. The function is a modified version of seewave::spectro(), so it takes the same arguments.

Value

Image files in jpeg format with spectrograms in the working directory, one for each sound file in X. It also returns the file path of the images invisibly.

Details

This function aims to simplify the evaluation of the alignment of test sound files from align_test_files(). The function creates a single spectrogram for each sound file (saved at dest.path). Spectrograms include the first few seconds of the sound files (controlled by duration), which is usually enough to tell the precision of the alignment. The plots include vertical lines denoting the start and end of each sound, as well as the sound ID (sound.id column in X). Note that no plot is created in the R graphic device.

References

Araya-Salas, M., Grabarczyk, E. E., Quiroz-Oliva, M., Garcia-Rodriguez, A., & Rico-Guevara, A. (2025). Quantifying degradation in animal acoustic signals with the R package baRulho. Methods in Ecology and Evolution, 00, 1-12. https://doi.org/10.1111/2041-210X.14481

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
  # load example data
  data("test_sounds_est")

  # plot (look into temporary working directory `tempdir()`)
  plot_aligned_sounds(X = test_sounds_est, dest.path = tempdir(), duration = 3, ovlp = 0)
}
#> The image files have been saved in the directory path '/tmp/RtmpLKxQlp'