Fix misaligned start and end of test sounds

realign_test_sounds fixes misaligned start and end of test sounds in an extended selection table using spectrographic cross-correlation

realign_test_sounds(X, parallel = 1, cores = getOption("mc.cores", 1), 
pb = getOption("pb", TRUE), hop.size = getOption("hop.size", 11.6), 
wl = getOption("wl", NULL), ovlp = getOption("ovlp", 90), wn = 'hanning')

Arguments

X
object of class 'extended_selection_table' created by the function selection_table from the warbleR package. The object must include the following additional columns: 'sound.id', 'bottom.freq' and 'top.freq'.

parallel
DEPRECATED. Use 'cores' instead.

cores
Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control if progress bar is shown. Default is TRUE.

hop.size
A numeric vector of length 1 specifying the time window duration (in ms). Default is 11.6 ms, which is equivalent to 512 wl for a 44.1 kHz sampling rate. Ignored if 'wl' is supplied.

wl
A numeric vector of length 1 specifying the window length of the spectrogram, default is NULL. If supplied, 'hop.size' is ignored.

ovlp
Numeric vector of length 1 specifying % of overlap between two consecutive windows, as in spectro. Default is 90. High values of ovlp slow down the function but produce more accurate results.

wn
A character vector of length 1 specifying the window name as in ftwindow.

Value

A data frame, or extended selection table similar to input data (depending on argument 'output') in which time parameters (columns 'start' and 'end') have been tailored to more closely match the start and end of the reference sound.

Details

This function uses spectrographic cross-correlation to align the position in time of sounds with regard to a reference sound. The sound recorded at the closest distance to the source is used as reference. Precise alignment is crucial for downstream measures of sound degradation. The function calls warbleR's cross_correlation internally to align sounds using cross-correlation. The output extended selection table contains the new start and end values after alignment.

References

Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2Clark, C.W., Marler, P. & Beeman K. (1987). Quantitative analysis of animal vocal phonology: an application to Swamp Sparrow song. Ethology. 76:101-115.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
# load example data
data("playback_est_unaligned")

# method 1
realign_test_sounds(X = playback_est_unaligned)
}

#> running cross-correlation (step 1 of 2):

#>   |                                                  | 0 % ~calculating    |+++++++++                                         | 17% ~00s            |+++++++++++++++++                                 | 33% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 83% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> finding peaks and aligning (step 2 out of 2)
#>   |                                                  | 0 % ~calculating    |+++++++++                                         | 17% ~00s            |+++++++++++++++++                                 | 33% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 83% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  

#> Object of class 'extended_selection_table'

#> * The output of the following call:

#> realign_test_sounds(X = playback_est_unaligned)

#> 
#> Contains: 
#> *  A selection table data frame with 8 row(s) and 8 columns:

#> |   |sound.files    | selec|  start|    end| bottom.freq| top.freq|

#> |:--|:--------------|-----:|------:|------:|-----------:|--------:|

#> |5  |01m_re-rec.wav |     4| 1.0500| 1.2445|      1.5441|  10.8946|

#> |6  |05m_re-rec.wav |     4| 1.0477| 1.2422|      1.5441|  10.8946|

#> |7  |10m_re-rec.wav |     4| 1.0419| 1.2363|      1.5441|  10.8946|

#> |8  |15m_re-rec.wav |     4| 1.0419| 1.2363|      1.5441|  10.8946|

#> |21 |01m_re-rec.wav |     2| 0.5500| 0.7001|      1.7436|  11.0500|

#> |22 |05m_re-rec.wav |     2| 0.5465| 0.6966|      1.7436|  11.0500|

#> ... 2 more column(s) (distance, sound.id)

#>  and 2 more row(s)

#> 
#> * 4 wave object(s) (as attributes): 

#> 01m_re-rec.wav05m_re-rec.wav10m_re-rec.wav15m_re-rec.wav

#> 
#> * A data frame (check.results) generated by check_sels() (as attribute)

#> 
#> Additional information:
#> * The selection table was created by song (see 'class_extended_selection_table')

#> * 1 sampling rate(s) (in kHz): 22.1

#> * 1 bit depth(s): 8

#> * Created by warbleR < 1.1.21