Measure attenuation as signal-to-noise ratio

snr_attenuation measures attenuation as signal-to-noise ratio of signals referenced in an extended selection table.

snr_attenuation(X, mar, parallel = 1, pb = TRUE, eq.dur = FALSE,
noise.ref = "adjacent", type = 1, bp = NULL, wl = 10, output = "est")

Arguments

X
object of class 'selection_table', 'extended_selection_table' created by the function selection_table from the warbleR package.

mar
numeric vector of length 1. Specifies the margins adjacent to the start and end points of selection over which to measure noise.

parallel
Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control if progress bar is shown. Default is TRUE.

eq.dur
Logical. Controls whether the noise segment that is measured has the same duration to that of the signal (if TRUE. Default is FALSE). If TRUE then 'mar' and 'noise.ref' arguments are ignored.

noise.ref
Character vector of length 1 to determined if a noise segment to be used for measuring ambient noise. Two options are available:

adjacent: measure ambient noise right before the signal (using argument 'mar' to define duration of noise segments). If several 'noise' selections by sound file are supplied, then the root mean square of the amplitude envelope will be averaged across those selections.

custom: measure noise segments referenced in the selection table (labeled as 'noise' in the 'signal.id' column). Those segments will be used to apply the same noise reference to all signals in a sound file. Therefore, at least one 'noise' selection for each sound file must be provided.

type
Numeric vector of length 1. Selects the formula to be used to calculate the signal-to-noise ratio (S = signal , N = background noise):

1: ratio of S amplitude envelope root mean square to N amplitude envelope root mean square (rms(env(S))/rms(env(N)))

2: ratio of the difference between S amplitude envelope root mean square and N amplitude envelope root mean square to N amplitude enveloperoot mean square ((rms(env(S)) - rms(env(N)))/rms(env(N)), as proposed by Dabelsteen et al (1993))

bp
Numeric vector of length 2 giving the lower and upper limits of a frequency bandpass filter (in kHz). Default is NULL.

wl
A numeric vector of length 1 specifying the window length of the spectrogram for applying bandpass. Default is 10. Ignored if bp = NULL. Note that lower values will increase time resolution, which is more important for amplitude ratio calculations.

output
Character vector of length 1 to determine if an extended selection table ('est', default) or a data frame ('data.frame') is returned.

Value

Data frame or extended selection table (depending on 'output' argument) similar to input data, but also includes a new column (snr.attenuation) with the signal-to-noise ratio values.

Details

Signal-to-noise ratio measures signal amplitude level in relation to ambient noise. A general margin in which ambient noise will be measured must be specified. Alternatively, a selection of ambient noise can be used as reference (see 'noise.ref' argument). When margins overlap with another acoustic signal nearby, the signal-to-noise ratio (SNR) will be inaccurate, so margin length should be carefully considered. Any SNR less than or equal to one suggests background noise is equal to or overpowering the acoustic signal. The 'signal.id' column must be used to indicate which signals belong to the same category (e.g. song-types). The function will then compare each signal type to the corresponding reference signal. Two methods for calculating signal-to-noise ratio are provided (see 'type' argument).

Examples

{
# load example data
data("playback_est")

# using noise reference selections 
snr_attenuation(X = playback_est, mar = 0.05, noise.ref = 'custom')

#' # remove noise selections
playback_est <- playback_est[playback_est$signal.id != "noise", ]
# using margin for noise of 0.05 and adjacent noise reference
snr_attenuation(X = playback_est, mar = 0.05, noise.ref = 'adjacent')
}

#>   |                                                  | 0 % ~calculating    |++                                                | 4 % ~00s            |++++                                              | 7 % ~00s            |++++++                                            | 11% ~00s            |++++++++                                          | 14% ~00s            |+++++++++                                         | 18% ~00s            |+++++++++++                                       | 21% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 29% ~00s            |+++++++++++++++++                                 | 32% ~00s            |++++++++++++++++++                                | 36% ~00s            |++++++++++++++++++++                              | 39% ~00s            |++++++++++++++++++++++                            | 43% ~00s            |++++++++++++++++++++++++                          | 46% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |+++++++++++++++++++++++++++                       | 54% ~00s            |+++++++++++++++++++++++++++++                     | 57% ~00s            |+++++++++++++++++++++++++++++++                   | 61% ~00s            |+++++++++++++++++++++++++++++++++                 | 64% ~00s            |++++++++++++++++++++++++++++++++++                | 68% ~00s            |++++++++++++++++++++++++++++++++++++              | 71% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 79% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 82% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 96% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#>   |                                                  | 0 % ~calculating    |+++                                               | 5 % ~00s            |+++++                                             | 10% ~00s            |++++++++                                          | 15% ~00s            |++++++++++                                        | 20% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++++                                | 35% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  

#> object of class 'extended_selection_table' 
#>  contains a selection table data frame with 20 rows and 9 columns: 
#>                                 sound.files selec start       end bottom.freq
#> 1 Playback_sound_file_LBH_degradation.wav_1     1  0.05 0.2072789    2.238685
#> 2                2W.5m.2019.7.1.10.10.wav_1     1  0.05 0.2072789    2.238685
#> 3               2W.10m.2019.7.1.10.20.wav_1     1  0.05 0.2072789    2.238685
#> 4               2W.15m.2019.7.1.10.30.wav_1     1  0.05 0.2072789    2.238685
#> 5 Playback_sound_file_LBH_degradation.wav_1     4  1.05 1.2444898    1.544095
#> 6                2W.5m.2019.7.1.10.10.wav_1     4  1.05 1.2444898    1.544095
#>   top.freq signal.id distance snr.attenuation
#> 1 11.58918    CCE-F1        1        43.98279
#> 2 11.58918    CCE-F1        5        12.28970
#> 3 11.58918    CCE-F1       10         9.99017
#> 4 11.58918    CCE-F1       15         5.26064
#> 5 10.89459    CCE-I2        1        42.78732
#> 6 10.89459    CCE-I2        5        20.99407
#> ... and 14 more rows 
#> 4 wave objects (as attributes): 
#> [1] "Playback_sound_file_LBH_degradation.wav_1"
#> [2] "2W.5m.2019.7.1.10.10.wav_1"               
#> [3] "2W.10m.2019.7.1.10.20.wav_1"              
#> [4] "2W.15m.2019.7.1.10.30.wav_1"              
#> and a data frame (check.results) generated by checkres() (as attribute) 
#> the selection table was created by song (see 'class_extended_selection_table')