Measure atmospheric attenuation and absorption of sound

atmospheric_attenuation measures atmospheric attenuation and atmospheric absorption.

Usage

atmospheric_attenuation(f, temp, RH, p = 101325, 
formula = 1, spi = NULL, dist = NULL)

Arguments

f

numeric vector of length 1 with frequency (in Hertz).

temp

numeric vector of length 1 with frequency (in Celsius).

RH

numeric vector of length 1 with relative humidity

p

numeric vector of length 1 with ambient pressure in Pa (standard: 101325, default).

formula

1 = Bazley 1976, 2 = Rossing 2007 (p. 116, see details).

spi

numeric vector of length 1 with the initial sound pressure in Pa. Required for calculating atmospheric absorption. Default is NULL.

dist

numeric vector of length 1 with distance (m) over which a sound propagates. Required for calculating atmospheric absorption. Default is NULL.

Value

Returns atmospheric attenuation (in dB/m) of sound based on supplied parameters. If 'spi' and 'dist' are supplied the function also returns atmospheric absorption (in dB).

Details

Calculate the atmospheric attenuation based on temperature, relative humidity, pressure and sound frequency. The function can applied to formulae based on:

1: default. As used by Bazley (1976), Sound absorption in air at frequencies up to 100 kHz. NPL acoustics report Ac 74.

2: as used by Rossing (2007), Handbook of Acoustics, Springer.

If 'spi' and 'dist' are supplied the function also returns the atmospheric absorption (in dB).

References

Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
# load example data
data("playback_est")

#' # remove ambient selections
playback_est <- playback_est[playback_est$signal.type != "ambient", ]

# measure atmospheric attenuation formula 1
atmospheric_attenuation(f = 20000, temp = 20, RH = 90, p = 88000, formula = 1)
}
#> $atmospheric.attenuation
#> [1] 0.3267505
#>