Fix misaligned start and end of test sounds

auto_realign fixes misaligned start and end of test sounds in an extended selection table using spectrographic cross-correlation

auto_realign(X, parallel = NULL, cores = getOption("mc.cores", 1),
pb = getOption("pb", TRUE), hop.size = getOption("hop.size", 11.6),
wl = getOption("wl", NULL), ovlp = getOption("ovlp", 90), wn = 'hanning')

Arguments

X

object of class 'extended_selection_table' created by the function selection_table from the warbleR package. The object must include the following additional columns: 'sound.id', 'bottom.freq' and 'top.freq'.

parallel

DEPRECATED. Use 'cores' instead.

cores

Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb

Logical argument to control if progress bar is shown. Default is TRUE.

hop.size

A numeric vector of length 1 specifying the time window duration (in ms). Default is 11.6 ms, which is equivalent to 512 wl for a 44.1 kHz sampling rate. Ignored if 'wl' is supplied.

wl

A numeric vector of length 1 specifying the window length of the spectrogram, default is NULL. If supplied, 'hop.size' is ignored.

ovlp

Numeric vector of length 1 specifying % of overlap between two consecutive windows, as in spectro. Default is 90. High values of ovlp slow down the function but produce more accurate results.

wn

A character vector of length 1 specifying the window name as in ftwindow.

Value

Object 'X' in which time parameters (columns 'start' and 'end') have been tailored to more closely match the start and end of the reference sound.

Details

This function uses spectrographic cross-correlation to align the position in time of sounds with regard to a reference sound. The sound recorded at the closest distance to the source is used as reference. Precise alignment is crucial for downstream measures of sound degradation. The function calls warbleR's cross_correlation internally to align sounds using cross-correlation. The output extended selection table contains the new start and end values after alignment.

References

Araya-Salas, M. (2020). baRulho: baRulho: quantifying degradation of (animal) acoustic signals in R. R package version 1.0.2Clark, C.W., Marler, P. & Beeman K. (1987). Quantitative analysis of animal vocal phonology: an application to Swamp Sparrow song. Ethology. 76:101-115.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
  # load example data
  data("test_sounds_est")

  # create "unaligned_test_sounds_est" by
  # adding noise to "test_sounds_est" start and end
  unaligned_test_sounds_est <- test_sounds_est
  set.seed(123)
  noise_time <- sample(c(0.005, -0.005, 0.006, -0.006, 0, 0.002, -0.002),
    nrow(unaligned_test_sounds_est),
    replace = TRUE
  )

  attr(unaligned_test_sounds_est, "check.res")$start <-
    unaligned_test_sounds_est$start <- unaligned_test_sounds_est$start + noise_time

  attr(unaligned_test_sounds_est, "check.res")$end <- unaligned_test_sounds_est$end <-
    unaligned_test_sounds_est$end + noise_time

  # re align

  auto_realign(X = unaligned_test_sounds_est)
}
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Object of class 'extended_selection_table'
#> * The output of the following call:
#> auto_realign(X = unaligned_test_sounds_est)
#> 
#> Contains: 
#> *  A selection table data frame with 25 row(s) and 9 columns:
#> |   |sound.files    | selec|  start|    end| bottom.freq| top.freq|
#> |:--|:--------------|-----:|------:|------:|-----------:|--------:|
#> |8  |10m_closed.wav |     1| 0.0480| 0.1980|      1.3333|   2.6667|
#> |9  |10m_closed.wav |     2| 1.3038| 1.5039|      7.8750|   8.8050|
#> |10 |10m_closed.wav |     3| 1.3032| 1.5032|      3.2080|   4.0690|
#> |11 |10m_closed.wav |     4| 1.3023| 1.5024|      0.4220|   1.2230|
#> |12 |10m_closed.wav |     5| 1.3027| 1.5027|      6.9050|   7.9170|
#> |13 |10m_open.wav   |     1| 0.0450| 0.1950|      1.3333|   2.6667|
#> ... 3 more column(s) (sound.id, transect, distance)
#>  and 19 more row(s)
#> 
#> * 6 wave object(s) (as attributes): 
#> master.wav10m_closed.wav10m_open.wav1m_open.wav30m_closed.wav30m_open.wav
#> 
#> * A data frame (check.results) generated by check_sels() (as attribute)
#> 
#> Additional information:
#> * The selection table was created by song (see 'class_extended_selection_table')
#> * 1 sampling rate(s) (in kHz): 22.05
#> * 1 bit depth(s): 16
#> * Created by warbleR 1.1.29