Measure detection distance of sound

detection_distance detection distance of sounds.

detection_distance(X, cores = getOption("mc.cores", 1), 
pb = getOption("pb", TRUE), output = "est", 
hop.size = getOption("hop.size", 11.6), wl = getOption("wl", NULL), 
path = getOption("sound.files.path", "."), spl = NULL, spl.cutoff, temp = 20, 
rh = 60, pa = 101325, hab.att.coef = 0.02, max.distance = 1000,
resolution = 0.1, subtract.bgn = TRUE, envelope = "abs", mar = NULL)

Arguments

X
Object of class 'data.frame', 'selection_table' or 'extended_selection_table' (the last 2 classes are created by the function selection_table from the warbleR package) with the reference to the sounds in the master sound file. Must contain the following columns: 1) "sound.files": name of the .wav files, 2) "selec": unique selection identifier (within a sound file), 3) "start": start time and 4) "end": end time of selections, 5) "bottom.freq": low frequency for bandpass, 6) "top.freq": high frequency for bandpass and 7) "sound.id": ID of sounds used to identify counterparts across distances. Each sound must have a unique ID within a distance.

cores
Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control if progress bar is shown. Default is TRUE.

output
Character vector of length 1 to determine if an extended selection table ('est', default) or a data frame ('data.frame').

hop.size
A numeric vector of length 1 specifying the time window duration (in ms). Default is 11.6 ms, which is equivalent to 512 wl for a 44.1 kHz sampling rate. Ignored if 'wl' is supplied.

wl
A numeric vector of length 1 specifying the window length of the spectrogram, default is NULL. If supplied, 'hop.size' is ignored.

path
Character string containing the directory path where the sound files are found. Only needed when 'X' is not an extended selection table.

spl
A numeric vector of length 1 specifying the sound pressure level of sounds. If not supplied then it will be measured from the sounds themselves.

spl.cutoff
A numeric vector of length 1 specifying the sound pressure level cutoff to define if the sound is no longer detected. Ideally it should be estimated based on the sound detection threshold of the species.

temp
Numeric vector of length 1 with frequency (in Celsius). Default is 20.

rh
Numeric vector of length 1 with relative humidity (in percentage). Default is 60.

pa
Numeric vector of length 1 with ambient pressure in Pa (standard: 101325, default). Used for Atmospheric attenuation.

hab.att.coef
Attenuation coefficient of the habitat (in dB/kHz/m).

max.distance
Numeric vector of length 1 with the maximum distance (in m) at which detection would be evaluated. Note that the function calculates the expected sound pressure level values along a vector of distances to find the distance at which the expected sound pressure level equates 'spl.cutoff'. Default is 1000 m.

resolution
NUmeric vector of length 1 with the distance resolution (in m) for estimated detection distance. Higher resolutions take longer to estimate. Default is 0.1 m.

subtract.bgn
Logical argument to control if SPL from background noise is excluded from the measured signal SPL. Default is FALSE.

envelope
Character string vector with the method to calculate amplitude envelopes (in which SPL is measured, used required if 'spl' is not supplied), as in env. Must be either 'abs' (absolute envelope, default) or 'hil' (Hilbert transformation).

mar
numeric vector of length 1. Specifies the margins adjacent to the start and end points of selection over which to measure background noise. This is required to subtract background noise sound pressure level (so only needed when 'subtract.bgn = TRUE').

Value

Returns the geometric, atmospheric and habitat attenuation (in dB) as well as the combined attenuation.

A data frame, or extended selection table similar to input data (depending on argument 'output'), but including two new columns (reference and detection.distance).

Details

The function computes the maximum distance at which a sound would be detected, which is calculated as the distance in which the sound pressure level (SPL) goes below the specified SPL cutoff ('spl.cutoff')). The function uses internally attenuation to estimate SPL at increasing values until it reaches the defined cutoff. The peak frequency (calculated on the power spectrum of the reference sound) of the reference sound for each sound ID is used as the carrier frequency for distance estimation. The sound recorded at the lowest distance is used as reference. This function assumes that all recordings have been made at the same recording volume.

References

Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2Clark, C.W., Marler, P. & Beeman K. (1987). Quantitative analysis of animal vocal phonology: an application to Swamp Sparrow song. Ethology. 76:101-115.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
# load example data
data("playback_est")

detection_distance(X = playback_est[playback_est$distance == 1,], spl.cutoff = 5, mar = 0.05)
}

#> Preparing data for analysis (step 1 out of 3):
#>   |                                                  | 0 % ~calculating    |++++++++                                          | 14% ~00s            |+++++++++++++++                                   | 29% ~00s            |++++++++++++++++++++++                            | 43% ~00s            |+++++++++++++++++++++++++++++                     | 57% ~00s            |++++++++++++++++++++++++++++++++++++              | 71% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> Computing sound pressure level and peak frequency (step 2 out of 3):
#>   |                                                  | 0 % ~calculating    |++++++++                                          | 14% ~00s            |+++++++++++++++                                   | 29% ~00s            |++++++++++++++++++++++                            | 43% ~00s            |+++++++++++++++++++++++++++++                     | 57% ~00s            |++++++++++++++++++++++++++++++++++++              | 71% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> Computing detection distance (step 3 out of 3):
#>   |                                                  | 0 % ~calculating    |++++++++                                          | 14% ~00s            |+++++++++++++++                                   | 29% ~00s            |++++++++++++++++++++++                            | 43% ~00s            |+++++++++++++++++++++++++++++                     | 57% ~00s            |++++++++++++++++++++++++++++++++++++              | 71% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  

#> Object of class 'extended_selection_table'

#> * The output of the following call:

#> detection_distance(X = playback_est[playback_est$distance ==  1, ], spl.cutoff = 5, mar = 0.05)

#> 
#> Contains: 
#> *  A selection table data frame with 7 row(s) and 10 columns:

#> |   |sound.files    | selec| start|    end| bottom.freq| top.freq|

#> |:--|:--------------|-----:|-----:|------:|-----------:|--------:|

#> |1  |01m_re-rec.wav |     1| 0.050| 0.2073|      2.2387|  11.5892|

#> |5  |01m_re-rec.wav |     4| 1.050| 1.2445|      1.5441|  10.8946|

#> |9  |01m_re-rec.wav |     5| 1.550| 1.6815|      2.3770|  11.7275|

#> |13 |01m_re-rec.wav |     3| 0.798| 0.9553|      2.2387|  11.5892|

#> |17 |01m_re-rec.wav |     6| 1.765| 1.9223|      2.2387|  11.5892|

#> |21 |01m_re-rec.wav |     2| 0.550| 0.7001|      1.7436|  11.0941|

#> ... 4 more column(s) (distance, sound.id, reference, detection.distance)

#>  and 1 more row(s)

#> 
#> * 1 wave object(s) (as attributes): 

#> 01m_re-rec.wav

#> 
#> * A data frame (check.results) generated by check_sels() (as attribute)

#> 
#> Additional information:
#> * The selection table was created by song (see 'class_extended_selection_table')

#> * 1 sampling rate(s) (in kHz): 44.1

#> * 1 bit depth(s): 8

#> * Created by warbleR < 1.1.21