Measure amplitude envelope correlation

envelope_correlation measures amplitude envelope correlation of signals referenced in an extended selection table.

envelope_correlation(X, parallel = 1, pb = TRUE, method = 1, cor.method = "pearson",
ssmooth = NULL, msmooth = NULL, output = "est")

Arguments

X
object of class 'selection_table', 'extended_selection_table' created by the function selection_table from the warbleR package.

parallel
Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing). If NULL (default) then the current working directory is used.

pb
Logical argument to control if progress bar is shown. Default is TRUE.

method
Numeric vector of length 1 to indicate the 'experimental design' to measure amplitude envelope correlation. Two methods are available:

1: compare all signals with their counterpart that was recorded at the closest distance to source (e.g. compare a signal recorded at 5m, 10m and 15m with its counterpart recorded at 1m). This is the default method.

2: compare all signals with their counterpart recorded at the distance immediately before (e.g. a signal recorded at 10m compared with the same signal recorded at 5m, then signal recorded at 15m compared with same signal recorded at 10m and so on).

cor.method
Character string indicating the correlation coefficient to be applied ("pearson", "spearman", or "kendall", see cor).

ssmooth
Numeric vector of length 1 to determine the length of the sliding window used for a sum smooth for amplitude envelope calculation (used internally by env).

msmooth
Numeric vector of length 2 to smooth the amplitude envelope with a mean sliding window for amplitude envelope calculation. The first element is the window length (in number of amplitude values) and the second one the window overlap (used internally by env).

output
Character vector of length 1 to determine if an extended selection table ('est', default) or a data frame ('data.frame') is returned.

Value

Data frame or extended selection table (depending on 'output' argument) similar to input data, but also includes a new column with the calculated amplitude envelope correlation coefficients.

Details

Amplitude envelope correlation measures the similarity of 2 signals in the time domain. The function measures the envelope correlation coefficients of signals in which a reference playback has been re-recorded at increasing distances. Values range from 1 (identical amplitude envelope, i.e. no degradation) to 0. The 'signal.id' column must be used to indicate the function to only compare signals belonging to the same category (e.g. song-types). The function will then compare each signal type to the corresponding reference signal. Two methods for calculating envelope correlation are provided (see 'method' argument).

References

Araya-Salas, M. (2019), baRulho: a R package to quantify habitat-induced degradation of (animal) acoustic signals. R package version 1.0.0Apol, C.A., Sturdy, C.B. & Proppe, D.S. (2017). Seasonal variability in habitat structure may have shaped acoustic signals and repertoires in the black-capped and boreal chickadees. Evol Ecol. 32:57-74.

Examples

{
# load example data
data("playback_est")

# remove noise selections
playback_est <- playback_est[playback_est$signal.id != "noise", ]

# method 1
envelope_correlation(X = playback_est)

# method 2
envelope_correlation(X = playback_est, method = 2)
}

#> calculating amplitude envelopes (step 1 of 2):
#>   |                                                  | 0 % ~calculating    |+++                                               | 5 % ~00s            |+++++                                             | 10% ~01s            |++++++++                                          | 15% ~01s            |++++++++++                                        | 20% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++++                                | 35% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> calculating envelope cross-correlations (step 2 of 2):
#>   |                                                  | 0 % ~calculating    |+++                                               | 5 % ~00s            |+++++                                             | 10% ~00s            |++++++++                                          | 15% ~00s            |++++++++++                                        | 20% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++++                                | 35% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> calculating amplitude envelopes (step 1 of 2):
#>   |                                                  | 0 % ~calculating    |+++                                               | 5 % ~00s            |+++++                                             | 10% ~00s            |++++++++                                          | 15% ~00s            |++++++++++                                        | 20% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++++                                | 35% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> calculating envelope cross-correlations (step 2 of 2):
#>   |                                                  | 0 % ~calculating    |+++                                               | 5 % ~00s            |+++++                                             | 10% ~00s            |++++++++                                          | 15% ~00s            |++++++++++                                        | 20% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++++                                | 35% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  

#> object of class 'extended_selection_table' 
#>  contains a selection table data frame with 20 rows and 9 columns: 
#>                                 sound.files selec start       end bottom.freq
#> 1 Playback_sound_file_LBH_degradation.wav_1     1  0.05 0.2072789    2.238685
#> 2                2W.5m.2019.7.1.10.10.wav_1     1  0.05 0.2072789    2.238685
#> 3               2W.10m.2019.7.1.10.20.wav_1     1  0.05 0.2072789    2.238685
#> 4               2W.15m.2019.7.1.10.30.wav_1     1  0.05 0.2072789    2.238685
#> 5 Playback_sound_file_LBH_degradation.wav_1     4  1.05 1.2444898    1.544095
#> 6                2W.5m.2019.7.1.10.10.wav_1     4  1.05 1.2444898    1.544095
#>   top.freq signal.id distance env.cor
#> 1 11.58918    CCE-F1        1      NA
#> 2 11.58918    CCE-F1        5       1
#> 3 11.58918    CCE-F1       10       1
#> 4 11.58918    CCE-F1       15       1
#> 5 10.89459    CCE-I2        1      NA
#> 6 10.89459    CCE-I2        5       1
#> ... and 14 more rows 
#> 4 wave objects (as attributes): 
#> [1] "Playback_sound_file_LBH_degradation.wav_1"
#> [2] "2W.5m.2019.7.1.10.10.wav_1"               
#> [3] "2W.10m.2019.7.1.10.20.wav_1"              
#> [4] "2W.15m.2019.7.1.10.30.wav_1"              
#> and a data frame (check.results) generated by checkres() (as attribute) 
#> the selection table was created by song (see 'class_extended_selection_table')