Measure full spectrum sound noise profiles

noise_profile Measure full spectrum sound pressure levels (i.e. noise profiles) in sound files or extended selection tables.

noise_profile(X = NULL, files = NULL, mar = NULL, 
noise.ref = "adjacent", parallel = 1, pb = TRUE, path = NULL,
bp = NULL, hop.size = 1, wl = NULL, PSD = FALSE, norm = TRUE, dB = "A", averaged = TRUE)

Arguments

X
object of class 'extended_selection_table' created by the function selection_table from the warbleR package. Default is NULL.

files
Character vector with names of wave files to be analyzed. Files must be found in 'path' supplied (or in the working directory if 'path' is not supplied). Default is NULL.

mar
numeric vector of length 1. Specifies the margins adjacent to the start and end points of selection over which to measure ambient noise. Required if 'X' is supplied and ignored if not supplied. Default is NULL.

noise.ref
Character vector of length 1 to determined which noise segment must be used for measuring ambient noise. Ignored if 'X' is not supplied. Two options are available:

adjacent: measure ambient noise right before the signal (using argument 'mar' to define duration of ambient noise segments).

custom: measure ambient noise segments referenced in the selection table (labeled as 'ambient' in the 'signal.type' column).

parallel
Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control if progress bar is shown. Default is TRUE.

path
Character string containing the directory path where the sound files are located. If NULL (default) then the current working directory is used.

bp
Numeric vector of length 2 giving the lower and upper limits of a frequency bandpass filter (in kHz). Default is NULL.

hop.size
A numeric vector of length 1 specifying the time window duration (in ms). Default is 1 ms, which is equivalent to ~45 wl for a 44.1 kHz sampling rate. Ignored if 'wl' is supplied.

wl
A numeric vector of length 1 specifying the window length of the spectrogram, default is NULL. Ignored if bp = NULL. If supplied, 'hop.size' is ignored. Note that lower values will increase time resolution, which is more important for amplitude ratio calculations.

PSD
Logical to control whether the Probability Mass Function (the probability distribution of frequencies). See meanspec. Default is FALSE.

norm
Logical to control whether amplitude values are normalized (divided by the maximum) so the highest value is 1. See meanspec. Default is TRUE.

dB
A character string of length 1 specifying the type dB to return: "max0" for a maximum dB value at 0, "A", "B", "C", "D", and "ITU" for common dB weights. See meanspec. Default is "A".

averaged
Logical to control if frequency spectra are averaged within a sound file. Default is TRUE.

Value

A list containing the the frequency spectra for each sound file or wave object (if 'X' is supplied).

Details

The function `noise_profile()` allows to estimate full spectrum sound pressure levels (i.e. noise profiles) of ambient noise. This can be done on extended selection tables (using the segments containing no signal) or over complete sound files in the working directory (or path supplied). The function uses meanspec internally to calculate frequency spectra.

References

Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
# load example data
data("playback_est")

# custom noise reference
noise_profile(X = playback_est, mar = 0.01, pb = FALSE, noise.ref = "custom")

# remove noise selections
pe <- playback_est[playback_est$signal.type != "ambient", ]
 
 noise_profile(X = pe, mar = 0.01, pb = FALSE, noise.ref = "adjacent")
}

#>       sound.files      freq        amp
#> 1  01m_re-rec.wav  1.002273  -5.721971
#> 2  05m_re-rec.wav  1.002273  -5.729903
#> 3  10m_re-rec.wav  1.002273  -5.720644
#> 4  15m_re-rec.wav  1.002273  -5.711130
#> 5  01m_re-rec.wav  2.004545 -40.506844
#> 6  05m_re-rec.wav  2.004545 -40.156630
#> 7  10m_re-rec.wav  2.004545 -39.413065
#> 8  15m_re-rec.wav  2.004545 -32.194383
#> 9  01m_re-rec.wav  3.006818 -49.094528
#> 10 05m_re-rec.wav  3.006818 -46.528108
#> 11 10m_re-rec.wav  3.006818 -46.618687
#> 12 15m_re-rec.wav  3.006818 -34.804503
#> 13 01m_re-rec.wav  4.009091 -54.690393
#> 14 05m_re-rec.wav  4.009091 -50.579048
#> 15 10m_re-rec.wav  4.009091 -48.417097
#> 16 15m_re-rec.wav  4.009091 -39.180214
#> 17 01m_re-rec.wav  5.011364 -59.602309
#> 18 05m_re-rec.wav  5.011364 -53.233228
#> 19 10m_re-rec.wav  5.011364 -51.911241
#> 20 15m_re-rec.wav  5.011364 -43.293286
#> 21 01m_re-rec.wav  6.013636 -63.182520
#> 22 05m_re-rec.wav  6.013636 -58.699244
#> 23 10m_re-rec.wav  6.013636 -54.656805
#> 24 15m_re-rec.wav  6.013636 -47.181135
#> 25 01m_re-rec.wav  7.015909 -65.666298
#> 26 05m_re-rec.wav  7.015909 -60.227954
#> 27 10m_re-rec.wav  7.015909 -56.814190
#> 28 15m_re-rec.wav  7.015909 -50.925429
#> 29 01m_re-rec.wav  8.018182 -69.290264
#> 30 05m_re-rec.wav  8.018182 -63.040709
#> 31 10m_re-rec.wav  8.018182 -57.386836
#> 32 15m_re-rec.wav  8.018182 -52.810468
#> 33 01m_re-rec.wav  9.020455 -69.570417
#> 34 05m_re-rec.wav  9.020455 -64.150070
#> 35 10m_re-rec.wav  9.020455 -58.996561
#> 36 15m_re-rec.wav  9.020455 -53.395612
#> 37 01m_re-rec.wav 10.022727 -73.044066
#> 38 05m_re-rec.wav 10.022727 -64.871929
#> 39 10m_re-rec.wav 10.022727 -61.148385
#> 40 15m_re-rec.wav 10.022727 -55.581374
#> 41 01m_re-rec.wav 11.025000 -73.765640
#> 42 05m_re-rec.wav 11.025000 -65.897917
#> 43 10m_re-rec.wav 11.025000 -62.707541
#> 44 15m_re-rec.wav 11.025000 -57.118786
#> 45 01m_re-rec.wav 12.027273 -75.417837
#> 46 05m_re-rec.wav 12.027273 -66.727830
#> 47 10m_re-rec.wav 12.027273 -64.208914
#> 48 15m_re-rec.wav 12.027273 -58.560034
#> 49 01m_re-rec.wav 13.029545 -75.942159
#> 50 05m_re-rec.wav 13.029545 -67.300412
#> 51 10m_re-rec.wav 13.029545 -64.932286
#> 52 15m_re-rec.wav 13.029545 -60.003674
#> 53 01m_re-rec.wav 14.031818 -77.009027
#> 54 05m_re-rec.wav 14.031818 -67.394874
#> 55 10m_re-rec.wav 14.031818 -65.040446
#> 56 15m_re-rec.wav 14.031818 -60.216267
#> 57 01m_re-rec.wav 15.034091 -77.979482
#> 58 05m_re-rec.wav 15.034091 -68.688015
#> 59 10m_re-rec.wav 15.034091 -66.032095
#> 60 15m_re-rec.wav 15.034091 -61.191970
#> 61 01m_re-rec.wav 16.036364 -78.333174
#> 62 05m_re-rec.wav 16.036364 -70.394644
#> 63 10m_re-rec.wav 16.036364 -66.434031
#> 64 15m_re-rec.wav 16.036364 -60.608169
#> 65 01m_re-rec.wav 17.038636 -80.091531
#> 66 05m_re-rec.wav 17.038636 -72.007798
#> 67 10m_re-rec.wav 17.038636 -66.899027
#> 68 15m_re-rec.wav 17.038636 -60.082887
#> 69 01m_re-rec.wav 18.040909 -80.002158
#> 70 05m_re-rec.wav 18.040909 -72.604253
#> 71 10m_re-rec.wav 18.040909 -67.137339
#> 72 15m_re-rec.wav 18.040909 -61.233618
#> 73 01m_re-rec.wav 19.043182 -81.011620
#> 74 05m_re-rec.wav 19.043182 -72.771296
#> 75 10m_re-rec.wav 19.043182 -68.339331
#> 76 15m_re-rec.wav 19.043182 -62.327506
#> 77 01m_re-rec.wav 20.045455 -82.010981
#> 78 05m_re-rec.wav 20.045455 -73.486001
#> 79 10m_re-rec.wav 20.045455 -68.183745
#> 80 15m_re-rec.wav 20.045455 -64.449787
#> 81 01m_re-rec.wav 21.047727 -82.754739
#> 82 05m_re-rec.wav 21.047727 -75.724601
#> 83 10m_re-rec.wav 21.047727 -70.677042
#> 84 15m_re-rec.wav 21.047727 -66.949034