Measure attenuation as signal-to-noise ratio

signal_to_noise_ratio measures attenuation as signal-to-noise ratio of sounds referenced in an extended selection table.

signal_to_noise_ratio(X, mar, parallel = 1, cores = getOption("mc.cores", 1), 
pb = getOption("pb", TRUE), eq.dur = FALSE, noise.ref = "adjacent", type = 1, 
bp = 'freq.range', output = "est", hop.size = getOption("hop.size", 1), 
wl = getOption("wl", NULL), ovlp = getOption("ovlp", 0), 
path = getOption("sound.files.path", "."))

Arguments

X
Object of class 'data.frame', 'selection_table' or 'extended_selection_table' (the last 2 classes are created by the function selection_table from the warbleR package) with the reference to the sounds in the master sound file. Must contain the following columns: 1) "sound.files": name of the .wav files, 2) "selec": unique selection identifier (within a sound file), 3) "start": start time and 4) "end": end time of selections, 5) "bottom.freq": low frequency for bandpass, 6) "top.freq": high frequency for bandpass and 7) "sound.id": ID of sounds used to identify counterparts across distances (only needed for "custom" noise reference, see "noise.ref" argument).

mar
numeric vector of length 1. Specifies the margins adjacent to the start and end points of selection over which to measure ambient noise.

parallel
DEPRECATED. Use 'cores' instead.

cores
Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control if progress bar is shown. Default is TRUE.

eq.dur
Logical. Controls whether the ambient noise segment that is measured has the same duration to that of the sound (if TRUE. Default is FALSE). If TRUE then 'mar' and 'noise.ref' arguments are ignored.

noise.ref
Character vector of length 1 to determined which noise segment must be used for measuring ambient noise. Two options are available:

adjacent: measure ambient noise right before the sound (using argument 'mar' to define duration of ambient noise segments).

custom: measure ambient noise segments referenced in the selection table (labeled as 'ambient' in the 'sound.id' column). Those segments will be used to apply the same ambient noise reference to all sounds in a sound file. Therefore, at least one 'ambient' selection for each sound file must be provided. If several 'ambient' selections by sound file are supplied, then the root mean square of the amplitude envelope will be averaged across those selections.

type
Numeric vector of length 1. Selects the formula to be used to calculate the signal-to-noise ratio (S = signal , N = background noise):

1: ratio of S amplitude envelope root mean square to N amplitude envelope root mean square (20 * log10(rms(env(S))/rms(env(N)))) as described by Darden (2008).

2: ratio of the difference between S amplitude envelope root mean square and N amplitude envelope root mean square to N amplitude envelope root mean square (20 * log10((rms(env(S)) - rms(env(N)))/rms(env(N))), as described by Dabelsteen et al (1993).

bp
Numeric vector of length 2 giving the lower and upper limits of a frequency bandpass filter (in kHz). Alternatively, when set to 'freq.range' (default) the function uses the 'bottom.freq' and 'top.freq' as the bandpass range.

output
Character vector of length 1 to determine if an extended selection table ('est', default) or a data frame ('data.frame').

hop.size
A numeric vector of length 1 specifying the time window duration (in ms). Default is 1 ms, which is equivalent to ~45 wl for a 44.1 kHz sampling rate. Ignored if 'wl' is supplied.

wl
A numeric vector of length 1 specifying the window length of the spectrogram, default is NULL. Ignored if bp = NULL. If supplied, 'hop.size' is ignored. Note that lower values will increase time resolution, which is more important for amplitude ratios calculations.

ovlp
Numeric vector of length 1 specifying the percent overlap between two consecutive windows, as in spectro. Default is 0. Only used for bandpass filtering.

path
Character string containing the directory path where the sound files are found. Only needed when 'X' is not an extended selection table.

Value

A data frame, or extended selection table similar to input data (depending on argument 'output'), but also includes a new column (signal.to.noise.ratio) with the signal-to-noise ratio values.

Details

Signal-to-noise ratio (SNR) measures sound amplitude level in relation to ambient noise. A general margin in which ambient noise will be measured must be specified. Alternatively, a selection of ambient noise can be used as reference (see 'noise.ref' argument). When margins overlap with another sound nearby, SNR will be inaccurate, so margin length should be carefully considered. Any SNR less than or equal to one suggests background noise is equal to or overpowering the sound. The function will measure signal-to-noise ratio within the supplied frequency range (e.g. bandpass) of the reference signal ('bottom.freq' and 'top.freq' columns in 'X') by default (that is, when bp = 'freq.range'.

References

Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2Dabelsteen, T., Larsen, O. N., & Pedersen, S. B. (1993). Habitat-induced degradation of sound signals: Quantifying the effects of communication sounds and bird location on blur ratio, excess attenuation, and signal-to-noise ratio in blackbird song. The Journal of the Acoustical Society of America, 93(4), 2206.Darden, SK, Pedersen SB, Larsen ON, & Dabelsteen T. (2008). Sound transmission at ground level in a short-grass prairie habitat and its implications for long-range communication in the swift fox *Vulpes velox*. The Journal of the Acoustical Society of America, 124(2), 758-766.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
# load example data
data("playback_est")

# using measure ambient noise reference selections
signal_to_noise_ratio(X = playback_est, mar = 0.05, noise.ref = 'custom')

# remove ambient selections
playback_est <- playback_est[playback_est$sound.id != "ambient", ]

# using margin for ambient noise of 0.05 and adjacent measure ambient noise reference
signal_to_noise_ratio(X = playback_est, mar = 0.05, noise.ref = 'adjacent')

}

#>   |                                                  | 0 % ~calculating    |++                                                | 4 % ~00s            |++++                                              | 7 % ~00s            |++++++                                            | 11% ~00s            |++++++++                                          | 14% ~00s            |+++++++++                                         | 18% ~00s            |+++++++++++                                       | 21% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 29% ~00s            |+++++++++++++++++                                 | 32% ~00s            |++++++++++++++++++                                | 36% ~00s            |++++++++++++++++++++                              | 39% ~00s            |++++++++++++++++++++++                            | 43% ~00s            |++++++++++++++++++++++++                          | 46% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |+++++++++++++++++++++++++++                       | 54% ~00s            |+++++++++++++++++++++++++++++                     | 57% ~00s            |+++++++++++++++++++++++++++++++                   | 61% ~00s            |+++++++++++++++++++++++++++++++++                 | 64% ~00s            |++++++++++++++++++++++++++++++++++                | 68% ~00s            |++++++++++++++++++++++++++++++++++++              | 71% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 79% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 82% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 96% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#>   |                                                  | 0 % ~calculating    |+++                                               | 5 % ~00s            |+++++                                             | 10% ~00s            |++++++++                                          | 15% ~00s            |++++++++++                                        | 20% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++++                                | 35% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  

#> Object of class 'extended_selection_table'

#> * The output of the following call:

#> signal_to_noise_ratio(X = playback_est, mar = 0.05, noise.ref = "adjacent")

#> 
#> Contains: 
#> *  A selection table data frame with 20 row(s) and 9 columns:

#> |sound.files    | selec|  start|    end| bottom.freq| top.freq|

#> |:--------------|-----:|------:|------:|-----------:|--------:|

#> |01m_re-rec.wav |     1| 0.0500| 0.2073|      2.2387|  11.5892|

#> |05m_re-rec.wav |     1| 0.0444| 0.2017|      2.2387|  11.5892|

#> |10m_re-rec.wav |     1| 0.0421| 0.1993|      2.2387|  11.5892|

#> |15m_re-rec.wav |     1| 0.0421| 0.1993|      2.2387|  11.5892|

#> |01m_re-rec.wav |     4| 1.0500| 1.2445|      1.5441|  10.8946|

#> |05m_re-rec.wav |     4| 1.0465| 1.2410|      1.5441|  10.8946|

#> ... 3 more column(s) (distance, sound.id, signal.to.noise.ratio)

#>  and 14 more row(s)

#> 
#> * 4 wave object(s) (as attributes): 

#> 01m_re-rec.wav05m_re-rec.wav10m_re-rec.wav15m_re-rec.wav

#> 
#> * A data frame (check.results) generated by check_sels() (as attribute)

#> 
#> Additional information:
#> * The selection table was created by song (see 'class_extended_selection_table')

#> * 1 sampling rate(s) (in kHz): 44.1

#> * 1 bit depth(s): 8

#> * Created by warbleR < 1.1.21