Measure spectrographic cross-correlation as a measure of sound distortion

spcc measures spectrographic cross-correlation as a measure of sound distortion in sounds referenced in an extended selection table.

spcc(X, parallel = 1, cores = 1, pb = TRUE,  method = 1, 
cor.method = "pearson", output = "est", 
hop.size = 11.6, wl = NULL, ovlp = 90, wn = 'hanning', path = NULL)

Arguments

X
Object of class 'data.frame', 'selection_table' or 'extended_selection_table' (the last 2 classes are created by the function selection_table from the warbleR package) with the reference to the sounds in the master sound file. Must contain the following columns: 1) "sound.files": name of the .wav files, 2) "selec": unique selection identifier (within a sound file), 3) "start": start time and 4) "end": end time of selections, 5) "bottom.freq": low frequency for bandpass, 6) "top.freq": high frequency for bandpass and 7) "sound.id": ID of sounds used to identify counterparts across distances. Each sound must have a unique ID within a distance.

parallel
DEPRECATED. Use 'cores' instead.

cores
Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control if progress bar is shown. Default is TRUE.

method
Numeric vector of length 1 to indicate the 'experimental design' for measuring envelope correlation. Two methods are available:

1: compare all sounds with their counterpart that was recorded at the closest distance to source (e.g. compare a sound recorded at 5m, 10m and 15m with its counterpart recorded at 1m). This is the default method.

2: compare all sounds with their counterpart recorded at the distance immediately before (e.g. a sound recorded at 10m compared with the same sound recorded at 5m, then sound recorded at 15m compared with same sound recorded at 10m and so on).

cor.method
Character string indicating the correlation coefficient to be applied ("pearson", "spearman", or "kendall", see cor).

output
Character vector of length 1 to determine if an extended selection table ('est', default) or a data frame ('data.frame').

hop.size
A numeric vector of length 1 specifying the time window duration (in ms). Default is 11.6 ms, which is equivalent to 512 wl for a 44.1 kHz sampling rate. Ignored if 'wl' is supplied.

wl
A numeric vector of length 1 specifying the window length of the spectrogram, default is NULL. If supplied, 'hop.size' is ignored.

ovlp
Numeric vector of length 1 specifying % of overlap between two consecutive windows, as in spectro. Default is 90. High values of ovlp slow down the function but produce more accurate results.

wn
A character vector of length 1 specifying the window name as in ftwindow.

path
Character string containing the directory path where the sound files are found. Only needed when 'X' is not an extended selection table.

Value

A data frame, or extended selection table similar to input data (depending on argument 'output'), but includes a new column (cross.correlation) with the spectrogram cross-correlation coefficients.

Details

Spectrographic cross-correlation measures frequency distortion of sounds as a similarity metric. Values close to 1 means very similar spectrograms (i.e. little sound distortion has occurred). Cross-correlation is measured of sounds in which a reference playback has been re-recorded at increasing distances. The 'sound.id' column must be used to indicate the function to only compare sounds belonging to the same category (e.g. song-types). The function compares each sound to the corresponding reference sound within the supplied frequency range (e.g. bandpass) of the reference sound ('bottom.freq' and 'top.freq' columns in 'X'). Two methods for calculating cross-correlation are provided (see 'method' argument). The function is a wrapper on warbleR's cross_correlation function.

References

Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2Clark, C.W., Marler, P. & Beeman K. (1987). Quantitative analysis of animal vocal phonology: an application to Swamp Sparrow song. Ethology. 76:101-115.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
# load example data
data("playback_est")

# method 1
spcc(X = playback_est, method = 1)

# method 2
spcc(X = playback_est, method = 2)
}

#> running cross-correlation (step 1 of 1):
#>   |                                                  | 0 % ~calculating    |++++                                              | 7 % ~00s            |+++++++                                           | 13% ~00s            |++++++++++                                        | 20% ~00s            |++++++++++++++                                    | 27% ~00s            |+++++++++++++++++                                 | 33% ~00s            |++++++++++++++++++++                              | 40% ~00s            |++++++++++++++++++++++++                          | 47% ~00s            |+++++++++++++++++++++++++++                       | 53% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |+++++++++++++++++++++++++++++++++++++             | 73% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> running cross-correlation (step 1 of 1):
#>   |                                                  | 0 % ~calculating    |++++                                              | 7 % ~00s            |+++++++                                           | 13% ~00s            |++++++++++                                        | 20% ~00s            |++++++++++++++                                    | 27% ~00s            |+++++++++++++++++                                 | 33% ~00s            |++++++++++++++++++++                              | 40% ~00s            |++++++++++++++++++++++++                          | 47% ~00s            |+++++++++++++++++++++++++++                       | 53% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |+++++++++++++++++++++++++++++++++++++             | 73% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  

#> Object of class 'extended_selection_table' 
#> * The output of the following call: 
#> spcc(X = playback_est, method = 2) 
#> 
#> Contains: 
#> *  A selection table data frame with 28 row(s) and 10 columns: 
#> |sound.files    | selec|  start|    end| bottom.freq| top.freq|
#> |:--------------|-----:|------:|------:|-----------:|--------:|
#> |01m_re-rec.wav |     1| 0.0500| 0.2073|      2.2387|  11.5892|
#> |05m_re-rec.wav |     1| 0.0444| 0.2017|      2.2387|  11.5892|
#> |10m_re-rec.wav |     1| 0.0421| 0.1993|      2.2387|  11.5892|
#> |15m_re-rec.wav |     1| 0.0421| 0.1993|      2.2387|  11.5892|
#> |01m_re-rec.wav |     4| 1.0500| 1.2445|      1.5441|  10.8946|
#> |05m_re-rec.wav |     4| 1.0465| 1.2410|      1.5441|  10.8946|
#> ... 4 more column(s) (distance, sound.id, reference, cross.correlation) and 22 more row(s) 
#> 
#> * 4 wave object(s) (as attributes): 01m_re-rec.wav 05m_re-rec.wav 10m_re-rec.wav 15m_re-rec.wav
#>  
#> * A data frame (check.results) generated by check_sels() (as attribute) 
#> 
#> Additional information:
#> * The selection table was created by song (see 'class_extended_selection_table') 
#> * 1 sampling rate(s) (in kHz): 44.1 
#> * 1 bit depth(s): 8 
#> * Created by warbleR < 1.1.21