Measure spectrographic cross-correlation as a measure of signal distortion

spcc measures spectrographic cross-correlation as a measure of signal distortion in signals referenced in an extended selection table.

spcc(X, parallel = 1, pb = TRUE,  method = 1,
cor.method = "pearson", hop.size = 11.6, wl = NULL, ovlp = 90, wn = 'hanning')

Arguments

X
object of class 'extended_selection_table' created by the function selection_table from the warbleR package. The object must include the following additional columns: 'signal.type', 'bottom.freq' and 'top.freq'.

parallel
Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control if progress bar is shown. Default is TRUE.

method
Numeric vector of length 1 to indicate the 'experimental design' for measuring envelope correlation. Two methods are available:

1: compare all signals with their counterpart that was recorded at the closest distance to source (e.g. compare a signal recorded at 5m, 10m and 15m with its counterpart recorded at 1m). This is the default method.

2: compare all signals with their counterpart recorded at the distance immediately before (e.g. a signal recorded at 10m compared with the same signal recorded at 5m, then signal recorded at 15m compared with same signal recorded at 10m and so on).

cor.method
Character string indicating the correlation coefficient to be applied ("pearson", "spearman", or "kendall", see cor).

hop.size
A numeric vector of length 1 specifying the time window duration (in ms). Default is 11.6 ms, which is equivalent to 512 wl for a 44.1 kHz sampling rate. Ignored if 'wl' is supplied.

wl
A numeric vector of length 1 specifying the window length of the spectrogram, default is NULL. If supplied, 'hop.size' is ignored.

ovlp
Numeric vector of length 1 specifying % of overlap between two consecutive windows, as in spectro. Default is 90. High values of ovlp slow down the function but produce more accurate results.

wn
A character vector of length 1 specifying the window name as in ftwindow.

Value

Extended selection table similar to input data, but includes a new column (cross.correlation) with the spectrogram cross-correlation coefficients.

Details

Spectrographic cross-correlation measures frequency distortion of signals as a similarity metric. Values close to 1 means very similar spectrograms (i.e. little signal distortion has occurred). Cross-correlation is measured of signals in which a reference playback has been re-recorded at increasing distances. The 'signal.type' column must be used to indicate the function to only compare signals belonging to the same category (e.g. song-types). The function compares each signal type to the corresponding reference signal within the supplied frequency range (e.g. bandpass) of the reference signal ('bottom.freq' and 'top.freq' columns in 'X'). Two methods for calculating cross-correlation are provided (see 'method' argument). The function is a wrapper on warbleR's xcorr function.

References

Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.0Clark, C.W., Marler, P. & Beeman K. (1987). Quantitative analysis of animal vocal phonology: an application to Swamp Sparrow song. Ethology. 76:101-115.

See also

Examples

{
# load example data
data("playback_est")

# method 1
spcc(X = playback_est, method = 1)

# method 2
spcc(X = playback_est, method = 2)
}

#> creating spectrogram matrices (step 1 of 2):
#>   |                                                  | 0 % ~calculating    |+++                                               | 5 % ~02s            |+++++                                             | 10% ~02s            |++++++++                                          | 15% ~01s            |++++++++++                                        | 20% ~01s            |+++++++++++++                                     | 25% ~01s            |+++++++++++++++                                   | 30% ~01s            |++++++++++++++++++                                | 35% ~01s            |++++++++++++++++++++                              | 40% ~01s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=01s  
#> running cross-correlation (step 2 of 2):
#>   |                                                  | 0 % ~calculating    |++++                                              | 7 % ~00s            |+++++++                                           | 13% ~00s            |++++++++++                                        | 20% ~00s            |++++++++++++++                                    | 27% ~00s            |+++++++++++++++++                                 | 33% ~00s            |++++++++++++++++++++                              | 40% ~00s            |++++++++++++++++++++++++                          | 47% ~00s            |+++++++++++++++++++++++++++                       | 53% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |+++++++++++++++++++++++++++++++++++++             | 73% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> creating spectrogram matrices (step 1 of 2):
#>   |                                                  | 0 % ~calculating    |+++                                               | 5 % ~00s            |+++++                                             | 10% ~00s            |++++++++                                          | 15% ~00s            |++++++++++                                        | 20% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++++                                | 35% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> running cross-correlation (step 2 of 2):
#>   |                                                  | 0 % ~calculating    |++++                                              | 7 % ~00s            |+++++++                                           | 13% ~00s            |++++++++++                                        | 20% ~00s            |++++++++++++++                                    | 27% ~00s            |+++++++++++++++++                                 | 33% ~00s            |++++++++++++++++++++                              | 40% ~00s            |++++++++++++++++++++++++                          | 47% ~00s            |+++++++++++++++++++++++++++                       | 53% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |+++++++++++++++++++++++++++++++++++++             | 73% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  

#> object of class 'extended_selection_table' 
#>  contains a selection table data frame with 28 rows and 10 columns: 
#>      sound.files selec      start       end bottom.freq top.freq signal.type
#> 1 01m_re-rec.wav     1 0.05000000 0.2072789    2.238685 11.58918      CCE-F1
#> 2 05m_re-rec.wav     1 0.04440598 0.2016849    2.238685 11.58918      CCE-F1
#> 3 10m_re-rec.wav     1 0.04206882 0.1993477    2.238685 11.58918      CCE-F1
#> 4 15m_re-rec.wav     1 0.04206882 0.1993477    2.238685 11.58918      CCE-F1
#> 5 01m_re-rec.wav     4 1.05000000 1.2444898    1.544095 10.89459      CCE-I2
#> 6 05m_re-rec.wav     4 1.04650916 1.2409990    1.544095 10.89459      CCE-I2
#>   distance        reference cross.correlation
#> 1        1             <NA>                NA
#> 2        5 01m_re-rec.wav-1         0.7885754
#> 3       10 05m_re-rec.wav-1         0.8680275
#> 4       15 10m_re-rec.wav-1         0.8181406
#> 5        1             <NA>                NA
#> 6        5 01m_re-rec.wav-4         0.7239949
#> ... and 22 more rows 
#> 4 wave objects (as attributes): 
#> [1] "01m_re-rec.wav" "05m_re-rec.wav" "10m_re-rec.wav" "15m_re-rec.wav"
#> and a data frame (check.results) generated by check_sels() (as attribute) 
#> the selection table was created by song (see 'class_extended_selection_table') 
#> 1 sampling rate(s): 44.1 
#> 1 bit depth(s): 8 
#> created by warbleR < 1.1.21