Align start and end of signal using spectrographic cross-correlation

spcc_align aligns start and end of signal in an extended selection table using spectrographic cross-correlation

spcc_align(X, parallel = 1, pb = TRUE, hop.size = 11.6, wl = NULL, ovlp = 90, 
wn = 'hanning')

Arguments

X
object of class 'extended_selection_table' created by the function selection_table from the warbleR package. The object must include the following additional columns: 'signal.type', 'bottom.freq' and 'top.freq'.

parallel
Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb
Logical argument to control if progress bar is shown. Default is TRUE.

hop.size
A numeric vector of length 1 specifying the time window duration (in ms). Default is 11.6 ms, which is equivalent to 512 wl for a 44.1 kHz sampling rate. Ignored if 'wl' is supplied.

wl
A numeric vector of length 1 specifying the window length of the spectrogram, default is NULL. If supplied, 'hop.size' is ignored.

ovlp
Numeric vector of length 1 specifying % of overlap between two consecutive windows, as in spectro. Default is 90. High values of ovlp slow down the function but produce more accurate results.

wn
A character vector of length 1 specifying the window name as in ftwindow.

Value

Extended selection table similar to input data in which time parameters (columns 'start' and 'end') have been tailored to more closely match the start and end of the reference signal.

Details

This function uses spectrographic cross-correlation to align the position in time of signals with regard to a reference signal. The signal recorded at the closest distance to the source is used as reference. Precise alignment is crucial for downstream measures of signal degradation. The function calls warbleR's xcorr and find_peaks internally to align signals using cross-correlation. The output extended selection table contains the new start and end values after alignment.

References

Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2Clark, C.W., Marler, P. & Beeman K. (1987). Quantitative analysis of animal vocal phonology: an application to Swamp Sparrow song. Ethology. 76:101-115.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

{
# load example data
data("playback_est_unaligned")

# method 1
spcc_align(X = playback_est_unaligned)
}

#> running cross-correlation (step 3 of 3):
#>   |                                                  | 0 % ~calculating    |+++++++++                                         | 17% ~00s            |+++++++++++++++++                                 | 33% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 83% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=01s  
#> finding peaks and aligning (step 3 of 3)
#>   |                                                  | 0 % ~calculating    |+++++++++                                         | 17% ~00s            |+++++++++++++++++                                 | 33% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 83% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  

#> Object of class 'extended_selection_table' 
#> Contains: 
#> * A selection table data frame with 8 rows and 8 columns: 
#>         sound.fi   selec   start    end   bottom.f   top.freq   signal.t  
#>         les                                    req              ype       
#>       ────────────────────────────────────────────────────────────────────
#>         01m_re-r       4   1.05    1.24       1.54       10.9   CCE-I2    
#>         ec.wav                                                            
#>         05m_re-r       4   1.96    2.16       1.54       10.9   CCE-I2    
#>         ec.wav                                                            
#>         10m_re-r       4   1.96    2.15       1.54       10.9   CCE-I2    
#>         ec.wav                                                            
#>         15m_re-r       4   1.96    2.15       1.54       10.9   CCE-I2    
#>         ec.wav                                                            
#>         01m_re-r       2   0.55    0.7        1.74       11.1   SUR-A4    
#>         ec.wav                                                            
#>         05m_re-r       2   0.991   1.14       1.74       11.1   SUR-A4    
#>         ec.wav                                                            
#> 7/8 columns shown.
#> ... and 2 more rows 
#> 
#> * 4 wave objects (as attributes): 01m_re-rec.wav 05m_re-rec.wav 10m_re-rec.wav 15m_re-rec.wav
#>  
#> * A data frame (check.results) generated by check_sels() (as attribute) 
#> 
#> Additional information:
#> * The selection table was created by song (see 'class_extended_selection_table') 
#> * 1 sampling rate(s) (in kHz): 22.1 
#> * 1 bit depth(s): 8 
#> * Created by warbleR < 1.1.21