Create synthetic sounds

synth_sounds create synthetic sounds

synth_sounds(steps = 10, am.amps = rep(c(1:4, 3:2), length.out = steps),
replicates = 1, frequencies, durations, nharmonics = 1, fm = FALSE, am = FALSE, 
mar = 0.05, seed = NULL, sig2 = 0.3, shuffle = FALSE,
hrm.freqs = c(1/2, 1/3, 2/3, 1/4, 3/4, 1/5, 1/6, 1/7, 1/8, 1/9, 1/10),
sampling.rate = 44.1)

Arguments

steps
Numeric vector of length 1. Controls the number of frequency steps in which each sound is split. Default is 10.

am.amps
Numeric vector with the relative amplitude for each step (see 'step' argument) to simulate amplitude modulation (only applied to the fundamental frequency). Should have the same length as the number of steps. Default is 1 (no amplitude modulation). The default value (rep(c(1:4, 3:2), length.out = steps)) has to amplitude peaks.

replicates
Numeric vector of length 1 indicating the number of replicates for each treatment combination. Default is 1. Useful for measuring variation in transmission parameters.

frequencies
Numeric vector with the different frequencies (in seconds) to synthesize.

durations
Numeric vector with the different durations (in seconds) to synthesize.

nharmonics
Numeric vector of length 1 specifying the number of harmonics to simulate. 1 indicates that only the fundamental frequency harmonic will be simulated.

fm
Logical to control if frequency modulated sounds are synthesize.

am
Logical to control if amplitude modulated sounds are synthesize.

mar
Numeric vector with the duration of margins of silence around sounds in seconds. Default is 0.05.

seed
Numeric vector of length 1. This allows users to get the same results in different runs (using set.seed internally). Default is NULL.

sig2
Numeric vector of length 1 defining the sigma value of the brownian motion model (used for simulating frequency modulation). Default is 0.3.

shuffle
Logical to control if the position of sounds is randomized. Useful to avoid having sounds from the same treatments next to each other. Defaul is FALSE.

hrm.freqs
Numeric vector with the frequencies of the harmonics relative to the fundamental frequency. The default values are c(1/2, 1/3, 2/3, 1/4, 3/4, 1/5, 1/6, 1/7, 1/8, 1/9, 1/10).

sampling.rate
Numeric vector of length 1. Sets the sampling frequency of the wave object (in kHz). Default is 44.1.

Value

A wave object containing the simulated songs. If 'selec.table' is TRUE the function saves the wave object as a '.wav' sound file in the working directory (or 'path') and returns a list including 1) a selection table with the start/end time, and bottom/top frequency of the sub-units and 2) the wave object.

Details

This functions uses a geometric (diff.fun == "GBM") or Brownian bridge (diff.fun == "BB") motion stochastic process to simulate modulation in animal vocalizations (i.e. frequency traces across time). The function can also simulate pure tones (diff.fun == "pure.tone", 'sig2' is ignored). Several song subunits (e.g. elements) can be simulated as well as the corresponding harmonics.

References

Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

if (FALSE) {
 # simulate a song with 3 elements and no harmonics
 sm_sng <- synth_sounds(n = 3, harms = 1)

 # plot spectro
 seewave::spectro(sm_sng)

 # simulate a song with 5 elements and 2 extra harmonics
sm_sng2 <- synth_sounds(n = 5, harms = 3)

 # plot spectrogram
 seewave::spectro(sm_sng2)

# six pure tones with frequency ranging form 4 to 6 and returning selection table
sm_sng <- synth_sounds(n = 6, harms = 1, seed = 1, diff.fun = "pure.tone",
                  freqs = seq(4, 6, length.out = 6), selec.table = TRUE,
                  path = tempdir())

# plot spectro
seewave::spectro(sm_sng$wave, flim = c(2, 8))

# selection table
sm_sng$selec.table
}